ed448

Module pkcs8

PKCS#8 private key support.

Implements Ed448 PKCS#8 private keys as described in RFC8410 Section 7: https://datatracker.ietf.org/doc/html/rfc8410#section-7

SemVer Notes

The pkcs8 module of this crate is exempted from SemVer as it uses a pre-1.0 dependency (the pkcs8 crate).

However, breaking changes to this module will be accompanied by a minor version bump.

Please lock to a specific minor version of the ed448 crate to avoid breaking changes when using this module.

Re-exports

pub use pkcs8::spki;

Structs

BitStringRef

ASN.1 BIT STRING type.

Document

ASN.1 DER-encoded document.

KeypairBytes

Ed448 keypair serialized as bytes.

ObjectIdentifier

Object identifier (OID).

OctetStringRef

ASN.1 OCTET STRING type: borrowed form.

PublicKeyBytes

Ed448 public key serialized as bytes.

SecretDocument

Secret Document type.

Enums

Error

Error type

KeyError

Key-related errors.

Constants

ALGORITHM_ID

Ed448 Algorithm Identifier.

ALGORITHM_OID

Algorithm ObjectIdentifier for the Ed448 digital signature algorithm (id-Ed448).

Traits

DecodePrivateKey

Parse a private key object from a PKCS#8 encoded document.

DecodePublicKey

Parse a public key object from an encoded SPKI document.

EncodePrivateKey

Serialize a private key object to a PKCS#8 encoded document.

EncodePublicKey

Serialize a public key object to a SPKI-encoded document.

Type Aliases

PrivateKeyInfoRef

PrivateKeyInfo with AnyRef algorithm parameters, and &[u8] key.

Result

Result type